Updated to yosys/CVC method

+++ js_cvc_interface.js
@@ -145,6 +145,20 @@ wss.on
                         })
 
                         ws.send ("Processing Verilog code...")
+                        // yosys_out = cp.execSync ('yosys -Q -T -q -p "synth_ice40 -top top -blif temp.blif" lab13.v 2>&1').toString()
+
+                        // error.push ()
+                        // for (var elm in yosys_out)
+                        //  {
+                        //     elm = yosys_out [elm]
+                        //     if (!elm.includes ("is used but has no driver") && elm.includes ('Warning'))
+                        //     {
+                        //         newelm = elm.replace ("Warning", "Error")
+                        //         lineno = parseInt (newelm.match (/ at lab13\.v\:([0-9]+)/)[1])
+                        //         newelm = 'Line ' + lineno.toString() + ': ' + newelm.replace (/ at lab13\.v\:([0-9]+)/, '').replace ('Error: ', '')
+                        //         console.log (newelm)
+                        //     }
+                        // }
                         cmd = 'cvc sim_modules/tb_ice40.sv tempcode/' + ws.unique_client + 
                               '/code.v -sv_lib sim_modules/svdpi.so -o tempcode/' + 
                               ws.unique_client + '/fpga'
@@ -181,26 +195,42 @@ wss.on
                             error.forEach (function (element) 
                                             { 
                                                 modded_err_rgx = /\*\*tempcode\/[a-z0-9]+\/code\.v\(([0-9]+)\)/
-                                                num = parseInt (element.match (modded_err_rgx) [1]) - (element.includes ("module does not exist") ? 28 : 29)
-                                                modded_err_msg = element.replace (modded_err_rgx, 'Line ' + num.toString() + ": ")
-                                                modded_err_msg = modded_err_msg.replace ("tempcode/" + ws.unique_client + '/code.v', 'your code ')
-                                                modded_error.push (modded_err_msg)
+                                                try {
+                                                    num = parseInt (element.match (modded_err_rgx) [1]) - (element.includes ("module does not exist") ? 28 : 29)
+                                                    modded_err_msg = element.replace (modded_err_rgx, 'Line ' + num.toString() + ": ")
+                                                    modded_err_msg = modded_err_msg.replace ("tempcode/" + ws.unique_client + '/code.v', 'your code ')
+                                                    modded_error.push (modded_err_msg)
+                                                }
+                                                catch (ex) {
+                                                    console.log ("Cannot parse this error: " + element)
+                                                    num = -1
+                                                }
                                             }
                                           )
-                            fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'code.v'));
-                            fs.rmdir (path.resolve (process.cwd(), 'tempcode', ws.unique_client), (err) => {
-                                if (err) { throw err; }
-                            })
-                            ws.send ("Compilation failed with the following error:\2" + modded_error.join ('\2'))
-                            ws.close()
+                            if (modded_error.length != '0')
+                            {
+                                fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'code.v'));
+                                fs.rmdir (path.resolve (process.cwd(), 'tempcode', ws.unique_client), (err) => {
+                                    if (err) { throw err; }
+                                })
+                                ws.send ("Compilation failed with the following error:\2" + modded_error.join ('\2'))
+                                ws.close()
+                            }
                         }
                         else
                         {
                             ws.currentState = "SIMULATE"
-                            console.log ("Starting " + ws.unique_client)
-                            var args = ('+interp sim_modules/tb_ice40.sv tempcode/' + ws.unique_client + '/code.v -sv_lib sim_modules/svdpi.so').split (" ")
                             var cp = require('child_process');
                             var env = Object.create( process.env );
+
+                            cp.execSync ('yosys -Q -T -q -p "synth_ice40 -top top -blif tempcode/' + ws.unique_client + 'temp.blif" tempcode/' + ws.unique_client + '/code.v 2>&1')
+                            cp.execSync ('yosys -o tempcode/' + ws.unique_client + '/struct_code.v tempcode/' + ws.unique_client + 'temp.blif')
+
+                            console.log ("Starting " + ws.unique_client)
+                            var args = ('+interp sim_modules/tb_struct_ice40.sv ' +
+                                         '/usr/local/bin/../share/yosys/ice40/cells_sim.v ' +
+                                         '/usr/local/bin/../share/yosys/ice40/cells_map.v ' +
+                                         'tempcode/' + ws.unique_client + '/struct_code.v -sv_lib sim_modules/svdpi.so').split (" ")
 
                             env.SVDPI_TO_PIPE='1';
                             env.SVDPI_FROM_PIPE='0';
@@ -213,7 +243,7 @@ wss.on
                             // console.log (running_simulations.size())
                             ws.send ("Simulation successfully started!")
 
-                            fs.writeFileSync (path.resolve (process.cwd(), 'logging', ws.unique_client, 'cvclog'), "CVC HAS STARTED ::: \n ::: ::: ::: ::: ::: \n", 'utf8', function (err) {
+                            fs.writeFileSync (path.resolve (process.cwd(), 'logging', ws.unique_client, 'cvclog'), "CVC HAS STARTED ::: \n", 'utf8', function (err) {
                                 if (err) { throw err; }
                             })
 
@@ -227,13 +257,13 @@ wss.on
                                     ws.send (msgdata)
                                 }
                                 catch (ex) {
-                                    ws.error_caught = true
                                     if (data.includes ("10 minutes exceeded"))
                                     {  
                                        ws.send ("TIME LIMIT EXCEEDED") 
                                     }
                                     else
                                     {
+                                        ws.error_caught = true
                                         fs.appendFile (path.resolve (process.cwd(), 'logging', ws.unique_client, 'cvclog'), data, 'utf8', function (err) {
                                             if (err) { throw err; }
                                         })
@@ -255,6 +285,8 @@ wss.on
                                 if (!ws.error_caught)
                                 {
                                     fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'code.v'));
+                                    fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'struct_code.v'));
+                                    // fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'temp.blif'));
                                     fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'fpga'));
                                     fs.rmdirSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client), (err) => {
                                         if (err) { throw err; }
@@ -283,6 +315,8 @@ wss.on
                                 if (!ws.error_caught)
                                 {
                                     fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'code.v'));
+                                    fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'struct_code.v'));
+                                    // fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'temp.blif'));
                                     fs.unlinkSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client, 'fpga'));
                                     fs.rmdirSync (path.resolve (process.cwd(), 'tempcode', ws.unique_client), (err) => {
                                         if (err) { throw err; }
